+++ clients/packages/checkout/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @polar-sh/checkout
+# @spaire/checkout
 
 ## 0.2.0
 
@@ -121,4 +121,4 @@
 
 ### Minor Changes
 
-- Initial release of @polar-sh/checkout
+- Initial release of @spaire/checkout